requirements.txt
bugswarm-common
git+https://github.com/dateutil/dateutil